Crate rust_info

rust_info

Extracts and provides the current rust compiler information.

This library main goal is to provide development/build tools such as cargo-makethe needed information on the current rust installation and setup.

Examples

fn main() {
    let rust_info = rust_info::get();

    println!("Version: {}", rust_info.version.unwrap());
    println!("Channel: {:#?}", rust_info.channel.unwrap());
    println!("Target Arch: {}", rust_info.target_arch.unwrap_or("unknown".to_string()));
    println!("Target Env: {}", rust_info.target_env.unwrap_or("unknown".to_string()));
    println!("Target OS: {}", rust_info.target_os.unwrap_or("unknown".to_string()));
    println!("Target Pointer Width: {}", rust_info.target_pointer_width.unwrap_or("unknown".to_string()));
    println!("Target Vendor: {}", rust_info.target_vendor.unwrap_or("unknown".to_string()));
    println!("Target Triple: {}", rust_info.target_triple.unwrap_or("unknown".to_string()));
}

Installation

In order to use this library, just add it as a dependency:

[dependencies]
rust_info = "*"

Contributing

See contributing guide

License

Developed by Sagie Gur-Ari and licensed under the Apache 2 open source license.

Modules

types

types

Structs

Options

Options used for querying rust info

Functions

get

Loads and returns the current rust compiler version and setup.
In case partial data is not available, those values will be set to Option::None.

get_with_options

Loads and returns the current rust compiler version and setup for a specified path.
In case partial data is not available, those values will be set to Option::None.