Function rust_info::get_with_options

pub fn get_with_options(options: &Options<'_>) -> RustInfo

Loads and returns the current rust compiler version and setup for a specified path.
In case partial data is not available, those values will be set to Option::None.

Example

fn main() {
    let opts = rust_info::Options::new().path(Some(std::path::Path::new("./").into()));
    let rust_info = rust_info::get_with_options(&opts);

    println!("Version: {}", rust_info.version.unwrap());
    println!("Channel: {:#?}", rust_info.channel.unwrap());
    println!("Target Arch: {}", rust_info.target_arch.unwrap_or("unknown".to_string()));
    println!("Target Env: {}", rust_info.target_env.unwrap_or("unknown".to_string()));
    println!("Target OS: {}", rust_info.target_os.unwrap_or("unknown".to_string()));
    println!("Target Pointer Width: {}", rust_info.target_pointer_width.unwrap_or("unknown".to_string()));
    println!("Target Vendor: {}", rust_info.target_vendor.unwrap_or("unknown".to_string()));
    println!("Target Triple: {}", rust_info.target_triple.unwrap_or("unknown".to_string()));
}